update Collection footers solution
