+++ Source/Table/Manager/BaseTableMahager+Headers.swift
@@ -1,0 +1,8 @@
+//
+//  BaseTableMahager+Headers.swift
+//  ReactiveDataDisplayManager
+//
+//  Created by Никита Коробейников on 18.06.2021.
+//
+
+import UIKit
