+++ Source/Table/Manager/BaseTableManager+Footers.swift
@@ -1,0 +1,8 @@
+//
+//  BaseTableManager+Footers.swift
+//  ReactiveDataDisplayManager
+//
+//  Created by Никита Коробейников on 18.06.2021.
+//
+
+import UIKit
